# doc/api.yaml
openapi: 3.0.0
info:
  description: WASAText app backend REST API.
  title: WASAText API
  version: 1.0.0
  contact:
    name: Francesco Calzona
    email: calzona.2046920@studenti.uniroma1.it
    url: www.placeholder.com
  license:
    name: ECL-2.0
    url: https://opensource.org/licenses/ECL-2.0
servers:
    - url: http://wasatext.com/v1
...
paths:
  /session:
    post:
      tags: ["login"]
      security: []
      summary: Logs in the user
      description: |-
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              type: object
              description: Username
              properties:
                name:
                  type: string
                  description: The user identifier
                  example: Maria
                  pattern: ^\w{3,16}$
                  minLength: 3
                  maxLength: 16
              example:
                UserExample:
                  properties:
                    name: Maria
        required: true
      responses:
        '201':
            description: User log-in action successful
            content:
              application/json:
                schema:
                  type: object
                  description: access token
                  properties:
                    identifier:
                      type: string
                      description: access token
                      pattern: ^[\w-$!&/(),.]{4,16}$
                      minLength: 4
                      maxLength: 16
                      example: "abcdef012345"
                  example:
                    UserToken:
                      properties:
                        identifier: alsdfkgnj34902857
        '400':
            description: Username parsed incorrectly or not valid
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
                example:
                  code: 400
                  message: The username is not valid
        '404':
            description: User not existing
            content:
              application/json:
                schema:
                  $ref : '#/components/schemas/Error'
                example:
                  code: 404
                  message: No user corresponds to the given username
        '500':
            description: Internal server error
            content:
              application/json:
                schema:
                    $ref : '#/components/schemas/Error'
                example:
                  code: 500
                  message: Internal server error
    put:
      tags: ["login"]
      security: []
      summary: Creates a user
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, an error will be returned.
      operationId: createUser
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              type: object
              description: Username
              properties:
                name:
                  type: string
                  description: The user identifier
                  example: Maria
                  pattern: ^\w{3,16}$
                  minLength: 3
                  maxLength: 16
              example:
                UserExample:
                  properties:
                    name: Maria
        required: true
      responses:
        '201':
            description: User creation action successful
            content:
              application/json:
                schema:
                  type: object
                  description: access token
                  properties:
                    identifier:
                      type: string
                      description: access token
                      pattern: ^[\w-$!&/(),.]{4,16}$
                      minLength: 4
                      maxLength: 16
                      example: "abcdef012345"
                  example:
                    UserToken:
                      properties:
                        identifier: alsdfkgnj34902857
        '400':
            description: Username parsed incorrectly or not valid
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
                example:
                  code: 400
                  message: The username is not valid
        '403':
            description: Username already in use
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
                example:
                  code: 403
                  message: The username is already in use
        '500':
            description: Internal server error
            content:
              application/json:
                schema:
                    $ref : '#/components/schemas/Error'
                example:
                  code: 500
                  message: Internal server error
...
  /settings/profilepicture:
    put:
      tags: ["user settings"]
      summary: Sets profile picture
      description: |-
        It sets the profile picture of the logged in user.
        This profile picture will be showed to other users.
      operationId: setMyPhoto
      requestBody:
        description: photo to be uploaded
        content:
          image/png:
            schema:
              description: photo
              type: string
              pattern: ^0b[01]{1,14}$
              minLength: 3
              maxLength: 16
              format: binary
            example: "0b00010010100"
        required: true
      responses:
        '202':
            description: Profile picture updated successfully
        '400':
            description: ProPic parsed incorrectly or not valid
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
                example:
                  code: 400
                  message: The picture is not valid
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
            description: Internal server error
            content:
              application/json:
                schema:
                    $ref : '#/components/schemas/Error'
                example:
                  code: 500
                  message: Internal server error
...
  /conversations/{conversationid}:
    parameters:
      - $ref: '#/components/parameters/conversationid'
    get:
      tags: ["conversations"]
      summary: Gets a conversation
      description: |-
        It retrieves a specific conversation (be it a private chat or a group for a specific user).
      operationId: getConversation
      responses:
        '200':
          description: Conversation obtained successfully
          content: 
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Chat'
                  - $ref: '#/components/schemas/Group'
              example:
                conversationid: 50
                user:
                  username: Fabio
                  propic: "0b00010010100"
                messages:
                  - messageid: 1
                    timestamp: "2017-07-21T17:32:28Z"
                    content: Hello
                    username: Mario
                    checkmarks: 1
                    comments: []
                    replyingto: 20
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                  $ref : '#/components/schemas/Error'
              example:
                code: 500
                message: Internal server error
    post:
      tags: ["messaging"]
      summary: Send a message
      description: |-
        It sends a message to a specific conversation (be it in a private chat or a group for a specific user).
      operationId: sendMessage
      requestBody:
        description: message to be sent
        content:
          application/json:
            schema:
                $ref: '#/components/schemas/Message'
            example:
                messageid: 1
                timestamp: "2017-07-21T17:32:28Z"
                content: Hello
                username: Mario
                checkmarks: 1
                comments: []
                replyingto: 20
        required: true
      responses:
        '204':
            description: Message sent
        '400':
            description: Message parsed incorrectly or not valid
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
                example:
                  code: 400
                  message: The message has one or more non-valid field(s)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/ConversationNotFound'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                  $ref : '#/components/schemas/Error'
              example:
                code: 500
                message: Internal server error
    delete:
      tags: ["groups"]
      summary: Leave a group
      description: |-
       If the current conversation is a group, it leaves the group. Otherwise, it throws an error.
      operationId: leaveGroup
      responses:
        '204':
          description: Group left successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403': 
          description: The current conversation is not a group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404': 
          $ref: '#/components/responses/ConversationNotFound'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                  $ref : '#/components/schemas/Error'
              example:
                code: 500
                message: Internal server error
...
components:
  schemas:
    Username:
      title: username
      description: This object represents a username
      type: string
      example: Maria
      pattern: ^\w{3,16}$
      minLength: 3
      maxLength: 16
    ConversationID:
      title: conversationid
      description: This object represents a conversation id
      type: integer
      example: 1
      minimum: 0
      maximum: 10000
    MessageID:
      title: message id
      description: This object represents a message id
      type: integer
      example: 1
      minimum: 0
      maximum: 10000
    CommentID:
      title: comment id
      description: This object represents a comment id
      type: integer
      example: 1
      minimum: 0
      maximum: 10000
    User:
      title: User
      description: This object represents a single user
      type: object
      properties:
        username:
          $ref: '#/components/schemas/Username'
        propic:
          type: string
          description: profile picture
          format: binary
          pattern: ^0b[01]{1,14}$
          minLength: 3
          maxLength: 16
          example: "0b00010010100"
      example:
        username: Fabio
        propic: "0b00010010100"
...
    Chat:
      title: Chat
      description: "This object represents a single private chat"
      type: object
      required:
        - messages
      properties:
        conversationid:
          $ref: '#/components/schemas/ConversationID'
        user:
          $ref: '#/components/schemas/User'
        messages:
          type: array
          description: Messages of the chat
          items:
              $ref: '#/components/schemas/Message'
          minItems: 1
          maxItems: 1000
          example:
            - messageid: 1
              timestamp: "2017-07-21T17:32:28Z"
              content: Hello
              username: Mario
              checkmarks: 1
              comments: []
              replyingto: 20
      example:
        conversationid: 50
        user:
          username: Fabio
          propic: "0b00010010100"
        messages:
          - messageid: 1
            timestamp: "2017-07-21T17:32:28Z"
            content: Hello
            username: Mario
            checkmarks: 1
            comments: []
            replyingto: 20
...
    Group:
      title: Group
      description: "This object represents a single group"
      type: object
      required:
        - members
        - messages
        - groupname
      properties:
        conversationid:
          $ref: '#/components/schemas/ConversationID'
        members:
          type: array
          description: group members
          items:
            $ref: '#/components/schemas/User'
          minItems: 2
          maxItems: 1000
          uniqueItems: true
          example:
            - username: Fabio
              propic: "0b00010010100"
            - username: Maria
              propic: "0b00010010110"
        messages:
          type: array
          description: Messages of the chat
          items:
            $ref: '#/components/schemas/Message'
          minItems: 1
          maxItems: 1000
          example:
            - messageid: 1
              timestamp: "2017-07-21T17:32:28Z"
              content: Hello
              username: Mario
              checkmarks: 1
              comments: []
              replyingto: 20
        groupname:
          type: string
          description: Name of the group
          pattern: ^\w{3,16}$
          minLength: 3
          maxLength: 16
          example: Family
        groupphoto:
          type: string
          description: Photo of the group
          pattern: ^[-A-Za-z0-9+/=]|=[^=]|={3,}$
          minLength: 3
          maxLength: 16
          format: byte
          example: Y2lhbw==
      example:
        conversationid: 10
        members:
          - username: Fabio
            propic: "0b00010010100"
          - username: Maria
            propic: "0b00010010110"
        messages: 
          - messageid: 1
            timestamp: "2017-07-21T17:32:28Z"
            content: Hello
            username: Mario
            checkmarks: 1
            comments: []
            replyingto: 20
        groupname: Family
        groupphoto: Y2lhbw==
    Message:
      title: Message
      description: "This object represents a single message"
      type: object
      required:
        - comments
      properties:
        messageid:
          $ref: '#/components/schemas/MessageID'
        timestamp:
          type: string
          description: Sent time of the message
          format: date-time
          pattern: ^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$
          minLength: 20
          maxLength: 20
          example: "2017-07-21T17:32:28Z"
        content:
          type: string
          description: Text of the message
          pattern: ^.{1,300}$
          minLength: 1
          maxLength: 300
          example: Hello
        photo:
          type: string
          description: A photo file in Base64 encoding.
          pattern: ^[-A-Za-z0-9+/=]|=[^=]|={3,16}$
          minLength: 3
          maxLength: 16
          format: byte
          example: Y2lhbw==
        username:
          $ref: '#/components/schemas/Username'     # sender's username
        checkmarks:
          type: number
          description: A single checkmark means that all recipients have received the message; a double one that everyone has read
          example: 1
          minimum: 0
          maximum: 2
        comments:
          type: array
          minItems: 0
          maxItems: 1000
          description: Comments to the message
          items:
            $ref: '#/components/schemas/Comment'
          uniqueItems: true
          example: []
        sentbyme:
          description: Boolean indicating whether the message was sent by myself
          type: boolean
          example: true
        replyingto:
          $ref: '#/components/schemas/MessageID'
      example:
        messageid: 1
        timestamp: "2017-07-21T17:32:28Z"
        content: Hello
        username: Mario
        checkmarks: 1
        comments: []
        replyingto: 20
    Comment:
      title: Comment
      description: "This object represents a single comment to a message"
      type: object
      required: 
        - reaction
      properties:
        commentid:
          $ref: '#/components/schemas/CommentID'
        sender:
          $ref: '#/components/schemas/Username'
        reaction:
          type: string
          description: Reaction emoji
          pattern: ^\w{3,9}$
          minLength: 3
          maxLength: 9
          enum: ["laugh", "sad", "thumbs_up", "surprised", "love", "pray"]
          example: sad
        sentbyme:
          description: Boolean indicating whether the message was sent by myself          
          type: boolean
          example: false
      example:
        commentid: 15
        sender: Gioia
        reaction: sad
        sentbyme: false
    Error:
      type: object
      description: A generic error message
      properties:
        code:
          description: HTTP error code
          type: number
          example: 401
        message:
          description: error message
          type: string
          pattern: ^.{1,300}$
          minLength: 1
          maxLength: 300
          example: UnauthorizedError
      required:
        - code
        - message
      example:
        code: 401
        message: UnauthorizedError
  parameters:
    conversationid:
      schema:
        $ref: '#/components/schemas/ConversationID'
      example: 1
      name: conversationid
      in: path
      description: ID of every conversation, associated both with the preview and the actual conversation
      required: true
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 401
            message: UnauthorizedError 
    ConversationNotFound:
        description: The conversation id given was not found
        content:
            application/json:
                schema:
                    $ref: '#/components/schemas/Error'
                example:
                  code: 404
                  message: UnauthorizedError
  securitySchemes:
    userID: # arbitrary name for the security scheme
      type: http
      description: The logging token of a single user
      scheme: bearer
      bearerFormat: JWT  # optional, for documentation purposes only
security:
  - userID: []